Move `messages` to `firezone-tunnel`
