+++ rust/connlib/tunnel/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0"
+base64 = { version = "0.22", default-features = false, features = ["std"] }
 bimap = "0.6"
 boringtun = { workspace = true }
 bytes = { version = "1.7", default-features = false, features = ["std"] }
@@ -13,8 +14,8 @@ connlib-shared = { workspace = true }
 derivative = "2.2.0"
 divan = { version = "0.1.14", optional = true }
 domain = { workspace = true }
-futures =  { version = "0.3", default-features = false, features = ["std", "async-await", "executor"] }
-futures-util =  { version = "0.3", default-features = false, features = ["std", "async-await", "async-await-macro"] }
+futures = { version = "0.3", default-features = false, features = ["std", "async-await", "executor"] }
+futures-util = { version = "0.3", default-features = false, features = ["std", "async-await", "async-await-macro"] }
 glob = "0.3.1"
 hex = "0.4.3"
 ip-packet = { workspace = true }
