+++ rust/connlib/tunnel/src/lib.rs
@@ -3,12 +3,12 @@
 //! This is both the wireguard and ICE implementation that should work in tandem.
 //! [Tunnel] is the main entry-point for this crate.
 
+use crate::messages::{Offer, Relay, ResolveRequest, SecretKey};
 use bimap::BiMap;
 use boringtun::x25519::StaticSecret;
 use chrono::Utc;
 use connlib_shared::{
-    messages::{ClientId, GatewayId, Offer, Relay, RelayId, ResolveRequest, ResourceId, SecretKey},
-    DomainName, PublicKey, ResourceView,
+    ClientId, DomainName, GatewayId, PublicKey, RelayId, ResourceId, ResourceView,
 };
 use io::Io;
 use ip_network::{Ipv4Network, Ipv6Network};
@@ -31,6 +31,7 @@ mod device_channel;
 mod dns;
 mod gateway;
 mod io;
+pub mod messages;
 mod peer;
 mod peer_store;
 #[cfg(all(test, feature = "proptest"))]
