+++ rust/connlib/tunnel/src/proptest.rs
@@ -1,10 +1,8 @@
-use connlib_shared::messages::{
-    client::{
-        ResourceDescription, ResourceDescriptionCidr, ResourceDescriptionDns,
-        ResourceDescriptionInternet, Site, SiteId,
-    },
-    ClientId, GatewayId, RelayId, ResourceId,
+use crate::messages::client::{
+    ResourceDescription, ResourceDescriptionCidr, ResourceDescriptionDns,
+    ResourceDescriptionInternet,
 };
+use connlib_shared::{ClientId, GatewayId, RelayId, ResourceId, Site, SiteId};
 use ip_network::{IpNetwork, Ipv4Network, Ipv6Network};
 use proptest::{
     arbitrary::{any, any_with},
