+++ rust/connlib/tunnel/src/tests/transition.rs
@@ -2,12 +2,11 @@ use crate::{
     client::{IPV4_RESOURCES, IPV6_RESOURCES},
     proptest::{host_v4, host_v6},
 };
+use connlib_shared::RelayId;
 
 use super::sim_net::{any_ip_stack, any_port, Host};
-use connlib_shared::{
-    messages::{client::ResourceDescription, DnsServer, RelayId, ResourceId},
-    DomainName,
-};
+use crate::messages::{client::ResourceDescription, DnsServer};
+use connlib_shared::{DomainName, ResourceId};
 use domain::base::Rtype;
 use prop::collection;
 use proptest::{prelude::*, sample};
