+++ rust/gateway/src/eventloop.rs
@@ -4,12 +4,12 @@ use crate::messages::{
 };
 use anyhow::Result;
 use boringtun::x25519::PublicKey;
-use connlib_shared::messages::{
-    ClientId, ConnectionAccepted, Interface, RelaysPresence, ResourceId,
-};
-use connlib_shared::{messages::GatewayResponse, DomainName};
+use connlib_shared::DomainName;
+use connlib_shared::{ClientId, ResourceId};
 #[cfg(not(target_os = "windows"))]
 use dns_lookup::{AddrInfoHints, AddrInfoIter, LookupError};
+use firezone_tunnel::messages::GatewayResponse;
+use firezone_tunnel::messages::{ConnectionAccepted, Interface, RelaysPresence};
 use firezone_tunnel::{DnsResourceNatEntry, GatewayTunnel};
 use futures::channel::mpsc;
 use futures_bounded::Timeout;
